--- a/Sources/Examples/RouteGuide/Model/route_guide.grpc.swift
+++ b/Sources/Examples/RouteGuide/Model/route_guide.grpc.swift
@@ -188,11 +188,11 @@ public protocol Routeguide_RouteGuideProvider: CallHandlerProvider {
 }
 
 extension Routeguide_RouteGuideProvider {
-  public var serviceName: String { return "routeguide.RouteGuide" }
+  public var serviceName: Substring { return "routeguide.RouteGuide" }
 
   /// Determines, calls and returns the appropriate request handler, depending on the request's method.
   /// Returns nil for methods not handled by this service.
-  public func handleMethod(_ methodName: String, callHandlerContext: CallHandlerContext) -> GRPCCallHandler? {
+  public func handleMethod(_ methodName: Substring, callHandlerContext: CallHandlerContext) -> GRPCCallHandler? {
     switch methodName {
     case "GetFeature":
       return CallHandlerFactory.makeUnary(callHandlerContext: callHandlerContext) { context in