--- a/dev/codegen-tests/05-service-only/golden/test.grpc.swift
+++ b/dev/codegen-tests/05-service-only/golden/test.grpc.swift
@@ -77,11 +77,11 @@ internal protocol Codegentest_FooProvider: CallHandlerProvider {
 }
 
 extension Codegentest_FooProvider {
-  internal var serviceName: String { return "codegentest.Foo" }
+  internal var serviceName: Substring { return "codegentest.Foo" }
 
   /// Determines, calls and returns the appropriate request handler, depending on the request's method.
   /// Returns nil for methods not handled by this service.
-  internal func handleMethod(_ methodName: String, callHandlerContext: CallHandlerContext) -> GRPCCallHandler? {
+  internal func handleMethod(_ methodName: Substring, callHandlerContext: CallHandlerContext) -> GRPCCallHandler? {
     switch methodName {
     case "Bar":
       return CallHandlerFactory.makeUnary(callHandlerContext: callHandlerContext) { context in