@Lukasa 2020-08-07
There's some noise in the way, but actually the win looks more like 5%. The results below are on a single thread, with the only difference being the application of this patch. The headline figure is an improvement of 5.45% on RPS.

Original:

```
Summary:
  Count:	171678
  Total:	30.00 s
  Slowest:	25.79 ms
  Fastest:	2.34 ms
  Average:	8.69 ms
  Requests/sec:	5722.51

Response time histogram:
  2.340 [1]	|
  4.685 [44]	|
  7.030 [32212]	|∎∎∎∎∎∎∎∎∎∎
  9.375 [125832]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  11.721 [2459]	|∎
  14.066 [4228]	|∎
  16.411 [6543]	|∎∎
  18.756 [184]	|
  21.101 [94]	|
  23.446 [12]	|
  25.791 [19]	|

Latency distribution:
  10 % in 6.89 ms 
  25 % in 8.47 ms 
  50 % in 8.61 ms 
  75 % in 8.71 ms 
  90 % in 8.85 ms 
  95 % in 13.84 ms 
  99 % in 15.66 ms 
```

With this patch:

```
Summary:
  Count:	181036
  Total:	30.00 s
  Slowest:	33.13 ms
  Fastest:	1.68 ms
  Average:	8.24 ms
  Requests/sec:	6034.45

Response time histogram:
  1.684 [1]	|
  4.829 [99]	|
  7.974 [40938]	|∎∎∎∎∎∎∎∎∎∎∎∎∎
  11.118 [128233]	|∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  14.263 [4777]	|∎
  17.408 [6687]	|∎∎
  20.552 [231]	|
  23.697 [10]	|
  26.842 [0]	|
  29.987 [3]	|
  33.131 [7]	|

Latency distribution:
  10 % in 6.53 ms 
  25 % in 8.02 ms 
  50 % in 8.15 ms 
  75 % in 8.25 ms 
  90 % in 8.39 ms 
  95 % in 13.11 ms 
  99 % in 14.83 ms 
```