@Lukasa 2020-08-07
There are two commits here: the first is the functional change, the second is the result of re-running codegen.